Move helper up
