+++ lib/responseDownloadFormats/helpers.ts
@@ -31,3 +31,10 @@ export const customTranslate = (translationFile: "my-forms" | "common") => {
 
   return { t };
 };
+
+export const getProperty = (field: string, lang: string): string => {
+  if (!field) {
+    throw new Error("Field does not exist");
+  }
+  return field + lang.charAt(0).toUpperCase() + lang.slice(1);
+};
