--- lib/responseDownloadFormats/html/helpers.ts
@@ -1,6 +1,0 @@
-export const getProperty = (field: string, lang: string): string => {
-  if (!field) {
-    throw new Error("Field does not exist");
-  }
-  return field + lang.charAt(0).toUpperCase() + lang.slice(1);
-};
